Fixing json parse error

+++ src/Solnet.Rpc/Converters/TransactionErrorJsonConverter.cs
@@ -105,7 +105,7 @@ namespace Solnet.Rpc.Converters
 
             if (reader.TokenType == JsonTokenType.Number)
             {
-                err.InstructionError.CustomError = reader.GetInt32();
+                err.InstructionError.CustomError = reader.GetInt64();
                 reader.Read(); //number
                 reader.Read(); //endobj
                 reader.Read(); //endarray
