+++ src/Solnet.Rpc/Models/InstructionError.cs
@@ -18,7 +18,7 @@
         /// <summary>
         /// Possible custom error id from a program.
         /// </summary>
-        public int? CustomError { get; set; }
+        public long? CustomError { get; set; }
 
         /// <summary>
         /// Possible string from borsh error.
